Tic Tac Toe Game using Minimax and Alpha-beta pruning
“TicTacToe” environment, the game will be
played between two computer agents: (a) rational agent: will always maximizes its utility and will
take the best move, (b) random agent: will take a move randomly between the “two best” moves,
that is, among the two moves with the lowest utility, it will make a move randomly. For leaf nodes
(when the tictactoe is filled), set the utility=1 if it is a win for the rational agent, utility = -1 if the
random agent wins, and utility=0 for a draw. We assume that the rational agent takes the first turn.
i. Minimax algorithm
ii. Alpha-beta pruning